Consolidate Tabs, tab-context, and TabSelector
