+++ src/components/Heading.js
@@ -5,8 +5,8 @@ import styled from '@emotion/styled';
 import { css } from '@emotion/react';
 import Loadable from '@loadable/component';
 import useScreenSize from '../hooks/useScreenSize';
-import TabSelectors from './TabSelectors';
-import { TabContext } from './tab-context';
+import TabSelectors from './Tabs/TabSelectors';
+import { TabContext } from './Tabs/tab-context';
 import ConditionalWrapper from './ConditionalWrapper';
 import Contents from './Contents/Contents';
 import Permalink from './Permalink';
