+++ src/components/Tabs/TabSelectors.js
@@ -1,9 +1,9 @@
 import React, { useContext, useMemo } from 'react';
 import { useTheme, css } from '@emotion/react';
 import { TabContext } from './tab-context';
-import Select from './Select';
-import { getPlaintext } from '../utils/get-plaintext';
-import { reportAnalytics } from '../utils/report-analytics';
+import Select from '../Select';
+import { getPlaintext } from '../../utils/get-plaintext';
+import { reportAnalytics } from '../../utils/report-analytics';
 
 const capitalizeFirstLetter = (str) => str.trim().replace(/^\w/, (c) => c.toUpperCase());
