+++ tests/unit/Tabs.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Tabs from '../../src/components/Tabs';
-import { TabProvider } from '../../src/components/tab-context';
+import Tabs from '../../src/components/Tabs/Tabs';
+import { TabProvider } from '../../src/components/Tabs/tab-context';
 import { theme } from '../../src/theme/docsTheme';
 
 // data for this component
